Fix PathMove Events pathnode coordinates.
To avoid making a sprite zigzagging along its path,
I turned the findPath function into using float values
for the pathnode returned.

This permitted to retain the starting offset of the sprite,
and try to keep it all along the way.

+++ src/modes/mode_bindings.cpp
@@ -505,8 +505,8 @@ void BindModeCode() {
 	module(hoa_script::ScriptManager->GetGlobalState(), "hoa_map")
 	[
 		class_<PathMoveSpriteEvent, SpriteEvent>("PathMoveSpriteEvent")
-			.def(constructor<std::string, uint32, int16, int16, bool>())
-			.def(constructor<std::string, VirtualSprite*, int16, int16, bool>())
+			.def(constructor<std::string, uint32, float, float, bool>())
+			.def(constructor<std::string, VirtualSprite*, float, float, bool>())
 			.def("SetRelativeDestination", &PathMoveSpriteEvent::SetRelativeDestination)
 			.def("SetDestination", &PathMoveSpriteEvent::SetDestination)
 	];
